AMP-4075
git-svn-id: file:///svn/trunk@7787 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/widget/action/ShowSectorByDonorChart.java
@@ -74,9 +74,9 @@ public class ShowSectorByDonorChart extends Action {
 		}else{
 			opt.setShowLabels(form.getShowLabel());
 		}
-		//HEIGHT
+		//HEIGHT		
 		if (form.getImageHeight() == null){
-        	opt.setHeight(new Integer(660));
+        	opt.setHeight(new Integer(350));
         }else{
         	opt.setHeight(form.getImageHeight());
         }
